Adds two new components with toggle functionality in specs
